Adjust import & exports to have cleaner paths. Move icons to asset folder
