+++ src/App.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Landing from "@landing";
+import { Landing } from "./pages";
 
 const App = () => {
   return (
