+++ src/components/common/index.tsx
@@ -1,0 +1,4 @@
+export * from "./button";
+export * from "./icon";
+export * from "./input";
+export * from "./modal";
