+++ src/components/common/modal/index.tsx
@@ -4,4 +4,4 @@ const Modal = () => {
   return <div>Modal</div>;
 };
 
-export default Modal;
+export { Modal };
