+++ src/components/index.tsx
@@ -1,0 +1,3 @@
+export * from "./common";
+export * from "./landing/signIn";
+export * from "./main";
