+++ src/components/main/index.tsx
@@ -1,9 +1,7 @@
-import React from 'react'
+import React from "react";
 
 const Main = () => {
-  return (
-    <div>Main</div>
-  )
-}
+  return <div>Main</div>;
+};
 
-export default Main
+export { Main };
